requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg==3.1.18
django-vite==3.0.4
python-dotenv==1.0.1
dj-database-url==2.1.0
